backend/requirements.txt
django-uniauth==1.2.0
Django==2.0.0
boto3==1.9.200
botocore==1.12.200
docutils==0.14
jmespath==0.9.4
python-dateutil==2.8.0
s3transfer==0.2.1
six==1.12.0
urllib3==1.25.3
PyJWT==1.7.1
